Support configuring dashboards on remote host
Previously, the dashboards were expected to be on the ansible control
host. Since the Grafana API may not always be accessible from there,
the dashboards should be copied to the remote host. This change
supports that.

An alternative to copying the contents of each dashboard back to the
ansible control host could be to compute the hash of each dashboard
on the remote host before comparing it to hashes of the existing
dashboards. However, to prevent overwriting tweaks to dashboards we
choose to compare them by title, and do not look for an exact
match. If an admin wants to re-install a particular dashboard they
can delete it, and then run this role. Since the dashboards are not
large this should improse little overhead.